Refactor Query-Input into API and add MentionController
