+++ API/src/main/java/de/brandwatch/minianalytics/api/APIApplication.java
@@ -1,7 +1,5 @@
 package de.brandwatch.minianalytics.api;
 
-import de.brandwatch.minianalytics.queryinput.solr.repository.MentionRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
@@ -11,8 +9,4 @@ public class APIApplication {
     public static void main(String[] args) {
         SpringApplication.run(APIApplication.class, args);
     }
-
-    @Autowired
-    MentionRepository mentionRepository;
-
 }
