+++ API/src/main/java/de/brandwatch/minianalytics/api/controller/MentionController.java
@@ -1,14 +1,17 @@
-package de.brandwatch.minianalytics.queryinput.controller;
+package de.brandwatch.minianalytics.api.controller;
 
-import de.brandwatch.minianalytics.queryinput.solr.model.Mention;
-import de.brandwatch.minianalytics.queryinput.solr.repository.MentionRepository;
+
+import com.google.gson.Gson;
+import de.brandwatch.minianalytics.api.solr.model.Mention;
+import de.brandwatch.minianalytics.api.solr.repository.MentionRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.time.Instant;
 import java.time.LocalDateTime;
@@ -18,19 +21,26 @@ import java.time.format.DateTimeFormatter;
 import java.time.temporal.TemporalAccessor;
 import java.util.List;
 
-@org.springframework.web.bind.annotation.RestController
-public class RestController {
+@RestController
+public class MentionController {
 
     private static final Logger logger = LoggerFactory.getLogger(RestController.class);
 
-    @Autowired
-    private MentionRepository mentionRepository;
+    private static final Gson gson = new Gson();
 
-    @RequestMapping(value = "/mentions/{queryID}")
+    private final MentionRepository mentionRepository;
+
+    @Autowired
+    public MentionController(MentionRepository mentionRepository) {
+        this.mentionRepository = mentionRepository;
+    }
+
+    @GetMapping(value = "/mentions/{queryID}")
     public ResponseEntity findMentionsFromQuery(@PathVariable String queryID, @RequestParam(value = "date", defaultValue = "") String date){
 
-        List<Mention> mentions = mentionRepository.findByQueryID(Long.parseLong(queryID));
+        logger.info("GET /mentions/" + queryID);
 
+        List<Mention> mentions = mentionRepository.findByQueryID(Long.parseLong(queryID));
 
         if(!date.equals("")){
             DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
@@ -45,9 +55,8 @@ public class RestController {
 
             logger.info("queryID: *?0* AND date: *?1*");
 
-            return ResponseEntity.status(200).body(mentionRepository.findMentionsAfterDate(Long.parseLong(queryID), dateBounds));
+            return ResponseEntity.status(200).body(gson.toJson(mentionRepository.findMentionsAfterDate(Long.parseLong(queryID), dateBounds)));
         }
-
-        return ResponseEntity.status(200).body(mentions);
+        return ResponseEntity.status(200).body(gson.toJson(mentions));
     }
 }
