+++ API/src/main/resources/application.properties
@@ -1,2 +1,17 @@
 spring.data.solr.host=http://127.0.0.1:8983/solr
 logging.level.org.springframework.data.solr.SolrTemplate=DEBUG
+
+## Connection to postgres db
+spring.datasource.url=jdbc:postgresql://localhost:5432/QueryDB
+spring.datasource.username=admin
+spring.datasource.password=admin
+
+spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults = false
+
+## Hibernate ddl
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.generate-ddl=true
+
+spring.datasource.driverClassName=org.postgresql.Driver
+
+spring.jpa.show-sql=true
