--- query-input/src/main/java/de/brandwatch/minianalytics/queryinput/controller/Controller.java
@@ -1,18 +1,0 @@
-package de.brandwatch.minianalytics.queryinput.controller;
-
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@org.springframework.stereotype.Controller
-public class Controller {
-
-    Logger logger = LoggerFactory.getLogger(Controller.class);
-
-    @RequestMapping(value = "/index")
-    public String index(){
-        return "index";
-    }
-
-}
