--- query-input/src/main/resources/templates/index.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset = "ISO-8859-1" />
-</head>
-<body>
-Get Mentions from Query ID: <input type="text" id="getMentionsFromQueryIDInput">
-<input type="button" onclick="getMentionsFromQueryID()" value="Get Mentions from Query">
-<br>
-<br>
-(Optional) since: <input type="date" id="datePicker">
-<br>
-<br>
-<textarea id="getMentionsFromQueryIDTextArea" rows="15" cols="80"></textarea>
-</body>
-<script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-<script>
-    function getMentionsFromQueryID() {
-        axios.get("http://localhost:8080/mentions/" + document.getElementById("getMentionsFromQueryIDInput").value, {
-            params: {
-                date: document.getElementById("datePicker").value
-            }
-        })
-            .then(function (response) {
-                document.getElementById("getMentionsFromQueryIDTextArea").innerHTML = JSON.stringify(response.data, null, '\t')
-            })
-
-        //console.log(document.getElementById("datePicker").value)
-    }
-</script>
-</html>
